added PPM and HPM compressors
darcs-hash:20061121152846-e5a07-3418ee367d4d7d303840af6eebf8ed4c20b3da95.gz

+++ net/contrapunctus/rngzip/util/BitOutputStream.java
@@ -111,7 +111,7 @@ public final class BitOutputStream implements Closeable
    /* Dump the contents of the bit buffer and reset it. */
    private void dumpByte() throws IOException
    {
-      out.write(buf);
+      out.write((byte)buf);
       buf = 0;
       pos = 0;
       /* invariants trivially hold here */
